excludes CUDA files and functions when using a non NVIDIA GPU machine

+++ src/FractalRenderer.h
@@ -3,7 +3,7 @@
 #include <string>
 #include "raylib.h"
 
-using Real = float;
+using Real = long double;
 
 class Window;
